% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{bici_routes}
\alias{bici_routes}
\title{Extract bike/bici routes for origins and destinations using CycleStreets via stplanr}
\usage{
bici_routes(
  od.data,
  distance.threshold = Inf,
  plan = c("quietest", "fastest", "balanced"),
  origin.col = names(od.data)[1],
  destination.col = names(od.data)[2],
  trips.col = names(od.data)[grep("trip", names(od.data))]
)
}
\arguments{
\item{od.data}{A sf object with OD data}

\item{distance.threshold}{distance threshold in metres to create a subset of the routes}

\item{plan}{one of "quietest" (default), "fastest", or "balanced" for cycle streets}

\item{origin.col}{string with name of the column with origin id}

\item{destination.col}{string with name of the column with destination id}

\item{trips.col}{string with name of the column with number of trips for each OD pair}
}
\value{
An sf object with the routes to school
}
\description{
Extract bike/bici routes for origins and destinations using CycleStreets via stplanr
}
\examples{
\dontrun{
bici_routes(od_data_almada)
}
}
