% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{bici_routes_osrm}
\alias{bici_routes_osrm}
\title{Extract bike/bici routes for origins and destinations}
\usage{
bici_routes_osrm(
  od.data,
  osrm.profile = "foot",
  origin.col = names(od.data)[1],
  destination.col = names(od.data)[2],
  trips.col = names(od.data)[grep("trip", names(od.data))]
)
}
\arguments{
\item{od.data}{A sf object with OD data}

\item{osrm.profile}{for the osrmRoute function}

\item{origin.col}{string with name of the column with origin id}

\item{destination.col}{string with name of the column with destination id}

\item{trips.col}{string with name of the column with number of trips for each OD pair}
}
\value{
An sf object with the routes to school
}
\description{
Extract bike/bici routes for origins and destinations
}
\examples{
\dontrun{
bici_routes(od.data)
}
}
