% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicischools-package.R
\docType{package}
\name{bicischools-package}
\alias{bicischools}
\alias{bicischools-package}
\title{bicischools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bicischools.github.io/bicischools/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
