% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bikebus_routes_almada}
\alias{bikebus_routes_almada}
\title{Bike bus routes in Almada a sample}
\format{
An sf object with X features and Y fields
}
\source{
\url{http://somewhere.com/}
}
\description{
A dataset containing bike bus routes to a school in Lisbon.
}
\examples{
bikebus_routes_almada
}
