% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{calc_stats}
\alias{calc_stats}
\title{Title}
\usage{
calc_stats(
  routes,
  rnet_plan,
  ordered_routes,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  min_trips = 0,
  buffer = 10
)
}
\arguments{
\item{routes}{}

\item{rnet_plan}{}

\item{ordered_routes}{}

\item{attribute_trips}{}

\item{min_trips}{}

\item{buffer}{}
}
\description{
Title
}
