% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_same_crs}
\alias{check_same_crs}
\title{Check CRS of two sf objects}
\usage{
check_same_crs(sf1, sf2)
}
\arguments{
\item{sf1}{An sf object.}

\item{sf2}{An sf object.}
}
\value{
TRUE if CRSs are the same, FALSE otherwise.
}
\description{
Check CRS of two sf objects
}
