% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_bus_routes.R
\name{cycle_bus_routes}
\alias{cycle_bus_routes}
\title{Title}
\usage{
cycle_bus_routes(
  routes,
  min_trips = 0,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  trips.col = names(routes)[grep("trip", names(routes))],
  buffer = 10
)
}
\arguments{
\item{routes}{}

\item{min_trips}{}

\item{attribute_trips}{}

\item{trips.col}{}

\item{buffer}{}
}
\value{
a dataset with cycle bus routes
}
\description{
Title
}
\examples{
\dontrun{
cycle_bus_routes(routes_almada)
}
}
