% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{filter_routes}
\alias{filter_routes}
\title{Filter bicycle routes based on spatial proximity and attribute ranking}
\usage{
filter_routes(
  routes,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  buffer = 300,
  top_n = 3
)
}
\arguments{
\item{routes}{An sf object containing route geometries with associated attributes.
Must contain a geometry column and the specified attribute ranking column.}

\item{attribute_trips}{Character string specifying which attribute to use for
ranking routes. One of "bicycle_godutch", "quietness", or "gradient_smooth".}

\item{buffer}{Numeric value specifying the minimum distance (in meters) between
route start points. Routes with start points closer than this distance will
be filtered out to avoid spatial overlap. Default is 300 meters.}

\item{top_n}{Integer specifying the maximum number of routes to return after
filtering and ranking. Default is 3.}
}
\value{
a dataset with the filtered routes
}
\description{
This function filters a set of bicycle routes by selecting routes that are
spatially separated (based on start point distances) and ranked by a specified
attribute. It's designed to identify optimal bike bus routes that don't overlap
at their origins and maximize cycling potential.
}
\examples{
\dontrun{
cycle_bus_routes <- cycle_bus_routes(routes_almada)
filter_routes(cycle_bus_routes)
}

}
\seealso{
\code{\link{cycle_bus_routes}} for creating route networks
}
