% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_models}
\alias{get_models}
\title{Get a model for SIM calculation}
\usage{
get_models(model)
}
\arguments{
\item{model}{A character string specifying the model name.}
}
\value{
A function representing the specified model.
}
\description{
Get a model for SIM calculation
}
\examples{
get_models("gravity")

}
