% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{match_centroids}
\alias{match_centroids}
\title{Title}
\usage{
match_centroids(
  routes_cents,
  top_routes,
  route_stats,
  origins,
  origin.id = names(origins)[1],
  oringin_route.id = "O",
  id.col = "id",
  dist_threshold = 500,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth")
)
}
\arguments{
\item{routes_cents}{}

\item{top_routes}{}

\item{route_stats}{}

\item{origins}{}

\item{origin.id}{}

\item{oringin_route.id}{}

\item{id.col}{}

\item{dist_threshold}{}

\item{attribute_trips}{}
}
\description{
Title
}
