% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{od_data_almada}
\alias{od_data_almada}
\title{Origin-destination data for Almada}
\format{
A data frame with X rows and Y variables
}
\source{
\url{http://somewhere.com/}
}
\description{
A dataset containing origin and destination data for the Almada region.
}
\examples{
od_data_almada
}
