% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{origins_almada}
\alias{origins_almada}
\title{Origin points in Lisbon}
\format{
An sf object with X features and Y fields
}
\source{
\url{http://somewhere.com/}
}
\description{
A dataset containing origin points within Lisbon.
}
\examples{
origins_almada
}
