% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{routes_almada}
\alias{routes_almada}
\title{Routes to school a sample}
\format{
An sf object with X features and Y fields
}
\source{
\url{http://somewhere.com/}
}
\description{
A dataset containing routes to a school in Lisbon.
}
\examples{
routes_almada
}
