% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{routes_network}
\alias{routes_network}
\title{Title}
\usage{
routes_network(
  routes,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  trips.col = names(routes)[grep("trip", names(routes))]
)
}
\arguments{
\item{routes}{}

\item{attribute_trips}{}

\item{trips.col}{}
}
\description{
Title
}
