% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{schools_lisbon}
\alias{schools_lisbon}
\title{School locations in Lisbon}
\format{
An sf object with X features and Y fields
}
\source{
\url{http://somewhere.com/}
}
\description{
A dataset containing school locations within Lisbon.
}
\examples{
schools_lisbon
}
