% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schools.R
\name{sim_schools}
\alias{sim_schools}
\title{Title}
\usage{
sim_schools(
  origins,
  destinations,
  model,
  ...,
  balancing = c("origins", "destinations", "both"),
  max_dist = Inf,
  origins.id = names(origins)[1],
  origins.size = names(origins)[2],
  destinations.id = names(destinations)[1],
  destinations.size = names(destinations)[2],
  keep_cols = FALSE,
  output_col = "modelled_trips",
  max_iter = 25
)
}
\arguments{
\item{origins}{An sf object representing the origins.}

\item{destinations}{An sf object representing the destinations.}

\item{model}{A character string specifying the model to use.}

\item{...}{Additional arguments to be passed to the model function.}

\item{balancing}{A character string specifying the balancing method. One of "origins", "destinations", or "both".}

\item{max_dist}{A numeric value specifying the maximum distance for interaction.}

\item{origins.id}{A character string specifying the ID column in the origins sf object.}

\item{origins.size}{A character string specifying the size column in the origins sf object.}

\item{destinations.id}{A character string specifying the ID column in the destinations sf object.}

\item{destinations.size}{A character string specifying the size column in the destinations sf object.}

\item{keep_cols}{A logical value. If TRUE, all columns from the input data are kept. If FALSE (default), only O, D and the output_col are kept.}

\item{output_col}{A character string specifying the name of the column for the modelled trips (default is "modelled_trips").}

\item{max_iter}{An integer, the maximum number of iterations for the balancing algorithm (default is 25).}
}
\value{
An sf object with the modelled trips.
}
\description{
Title
}
\examples{
# sim_schools(origins, destinations, model = "gravity")


}
