% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{batch_osrmRoutes}
\alias{batch_osrmRoutes}
\title{A function to query routes for several destinations to a single origin}
\usage{
batch_osrmRoutes(origins, destination, osrm.profile)
}
\arguments{
\item{origins}{}

\item{destination}{}

\item{osrm.profile}{}
}
\value{
an sf object with all the routes
}
\description{
A function to query routes for several destinations to a single origin
}
