% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{filter_routes}
\alias{filter_routes}
\title{Title}
\usage{
filter_routes(
  routes,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  buffer = 300,
  top_n = 3
)
}
\arguments{
\item{routes}{}

\item{attribute_trips}{}

\item{buffer}{}

\item{top_n}{}
}
\value{
a dataset with the filtered routes
}
\description{
Title
}
\examples{
\dontrun{
cycle_bus_routes(routes_almada)
}
}
