% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{match_centroids}
\alias{match_centroids}
\title{Title}
\usage{
match_centroids(
  routes_cents,
  top_routes,
  route_stats,
  centroids,
  id.col = names(routes_cents)[1],
  dist_threshold = 15,
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth")
)
}
\arguments{
\item{routes_cents}{}

\item{top_routes}{}

\item{route_stats}{}

\item{centroids_5km}{}
}
\description{
Title
}
