% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{prepare_od}
\alias{prepare_od}
\title{A function to get a nested OD data subset by school}
\usage{
prepare_od(od_data, trips.col, distance.threshold)
}
\arguments{
\item{od_data}{}

\item{trips.col}{}

\item{distance.threshold}{}
}
\value{
a nested tibble
}
\description{
A function to get a nested OD data subset by school
}
