% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{routes_pct_uptake}
\alias{routes_pct_uptake}
\title{Title}
\usage{
routes_pct_uptake(
  routes,
  trips.col = names(routes)[grep("trip", names(routes))],
  length.col = "length",
  max.length = 30000,
  hilliness.col = "route_hilliness"
)
}
\arguments{
\item{routes}{}

\item{trips.col}{}

\item{length.col}{}

\item{max.length}{}

\item{hilliness.col}{}
}
\description{
Title
}
