% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes_processing.R
\name{summarise_routes}
\alias{summarise_routes}
\title{Title}
\usage{
summarise_routes(
  routes,
  id_route.col = "route_number",
  origin.col = "O",
  destination.col = "D",
  trips.col = names(routes)[grep("trip", names(routes))],
  attribute_trips = c("bicycle_godutch", "quietness", "gradient_smooth"),
  length.col = "length"
)
}
\arguments{
\item{routes}{}

\item{id_route.col}{}

\item{origin.col}{}

\item{destination.col}{}

\item{trips.col}{}

\item{attribute_trips}{}

\item{length.col}{}
}
\description{
Title
}
